shifted slugcat stats to a separate file, removed extra debug
